Community/Fraud_Detection/requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
scikit-learn
joblib
python-dotenv
pandas
requests
